a-a-d-ai/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
langgraph==0.2.28
langchain==0.3.0
langchain-openai==0.2.0
openai==1.51.0
psycopg2-binary==2.9.9
pandas==2.2.2
plotly==5.24.0
python-dotenv==1.0.1
